Conflicts:
	gremlin-server/src/test/resources/org/apache/tinkerpop/gremlin/driver/remote/gremlin-server-integration.yaml
	gremlin-server/src/test/resources/org/apache/tinkerpop/gremlin/server/gremlin-server-integration.yaml
	gremlin-server/src/test/resources/org/apache/tinkerpop/gremlin/server/gremlin-server-performance.yaml
	gremlin-server/src/test/scripts/test-server-start.groovy
	gremlin-test/src/main/java/org/apache/tinkerpop/gremlin/process/traversal/step/branch/OptionalTest.java

+++ gremlin-dotnet/test/Gremlin.Net.IntegrationTest/Process/Traversal/DriverRemoteConnection/GraphTraversalTests.cs
@@ -55,8 +55,8 @@ namespace Gremlin.Net.IntegrationTest.Process.Traversal.DriverRemoteConnection
 
             var vertex = g.V(1).Next();
 
-            Assert.Equal(new Vertex((long) 1), vertex);
-            Assert.Equal((long) 1, vertex.Id);
+            Assert.Equal(new Vertex(1), vertex);
+            Assert.Equal(1, vertex.Id);
         }
 
         [Fact]
@@ -68,7 +68,7 @@ namespace Gremlin.Net.IntegrationTest.Process.Traversal.DriverRemoteConnection
 
             var traverser = g.V(1).NextTraverser();
 
-            Assert.Equal(new Traverser(new Vertex((long)1)), traverser);
+            Assert.Equal(new Traverser(new Vertex(1)), traverser);
         }
 
         [Fact]
@@ -138,7 +138,7 @@ namespace Gremlin.Net.IntegrationTest.Process.Traversal.DriverRemoteConnection
                 g.V(5).Repeat(__.Both().SimplePath()).Until(__.HasId(6)).Limit<Vertex>(1).Path().Next();
 
             Assert.Equal(4, shortestPath.Count);
-            Assert.Equal(new Vertex((long) 6), shortestPath[3]);
+            Assert.Equal(new Vertex(6), shortestPath[3]);
         }
 
         [Fact]
